voz_ia/requirements.txt
streamlit
speechrecognition
# elevenlabs
google-generativeai
pyaudio